Move notification setting out and reuse it for spaces
